requirements.txt
fastapi
uvicorn
streamlit
psycopg2-binary
sqlite-utils
qdrant-client
requests
matplotlib
pandas
openpyxl
xlrd